+++ frontend/package.json
@@ -11,7 +11,14 @@
   },
   "dependencies": {
     "react": "^18.3.1",
-    "react-dom": "^18.3.1"
+    "react-dom": "^18.3.1",
+    "cra-template": "1.2.0",
+    "lucide-react": "^0.474.0",
+    "react-router-dom": "^7.1.5",
+    "react-scripts": "5.0.1",
+    "react-signature-canvas": "^1.0.7",
+    "recharts": "^2.15.1",
+    "web-vitals": "^4.2.4"
   },
   "devDependencies": {
     "@eslint/js": "^9.17.0",
