+++ frontend/src/components/RegistrationInformation/RegistrationInformation.jsx
