--- frontend_temp/README.md
@@ -1,2 +1,0 @@
-# Health_Bridge
-A virtual platform for mothers and infants to record their health reports
